W-15633184: parse common avro fields in base parser
