+++ amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.jsonld
@@ -29,7 +29,7 @@
             ],
             "http://a.ml/vocabularies/shapes#range": [
               {
-                "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/long",
+                "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/value",
                 "@type": [
                   "http://a.ml/vocabularies/shapes#ScalarShape",
                   "http://a.ml/vocabularies/shapes#AnyShape",
@@ -44,21 +44,21 @@
                 ],
                 "http://www.w3.org/ns/shacl#name": [
                   {
-                    "@value": "long"
+                    "@value": "value"
                   }
                 ],
                 "http://a.ml/vocabularies/document-source-maps#sources": [
                   {
-                    "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/long/source-map",
+                    "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/value/source-map",
                     "@type": [
                       "http://a.ml/vocabularies/document-source-maps#SourceMap"
                     ],
                     "http://a.ml/vocabularies/document-source-maps#avro-schema": [
                       {
-                        "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/long/source-map/avro-schema/element_0",
+                        "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/value/source-map/avro-schema/element_0",
                         "http://a.ml/vocabularies/document-source-maps#element": [
                           {
-                            "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/long"
+                            "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field/scalar/value"
                           }
                         ],
                         "http://a.ml/vocabularies/document-source-maps#value": [
@@ -112,8 +112,9 @@
                     ]
                   },
                   {
-                    "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field_1/union/union/anyOf/any/default-any",
+                    "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/record.json#/shape/LongList/property/property/field_1/union/union/anyOf/scalar/default-scalar",
                     "@type": [
+                      "http://a.ml/vocabularies/shapes#ScalarShape",
                       "http://a.ml/vocabularies/shapes#AnyShape",
                       "http://www.w3.org/ns/shacl#Shape",
                       "http://a.ml/vocabularies/shapes#Shape",
